The Code IPNYB
This is the Jupyter Notebook File that has the structured code for the master's thesis implementation of the Vision Transformers to accurately classify the different morphologies of Galaxies using a simplified version of the famous Galaxy Zoo dataset- "Galaxy10Decals". This vision transformers are also compared with a traditional CNN model as well with comparison metrics being compared side by side for clearer understanding. The pre-processing and network implementation for the neural networks are all standardised and set in such a way that the comparison results have credibility.